BUG: Decref subspace when subspace is 0-d in fancy indexing
Not decrefing it causes a major memory leak. This adapts the
automated tests to cover this case as well and adds refcounting
to it. Replaces some 'skip' with 0s, but I believe the skip
tests are not that important. Added 1-d special case tests
using the same indices as the multi-dim test though.

Thanks to Julian Taylor for noting the bug.