MAINT: Remove some commented out code in test_indexing.py.
The code was intended to check refcounting, and indeed, exposed a long
standing bug by failing, which is why it was commented out. That bug is
now fixed and tested by gh-3631, so the commented out code could
theoretically be made to work, but that requires enough del this and del
the other that I think it is too fragile to keep. So delete it.