BUG: Catch DeprecationWarning in numpy/core/tests/test_numeric.py.
TestCreationFuncs.check_function was spewing Deprecation warnings in
release mode. AFAICT, it did not raise errors during development because
it was buried by a TypeError. The Deprecation in question is at line
1141 in multiarray/conversion_utils.c. Hiding the deprecation should
be harmless here, but when the Deprecation is turned into an error
it will need to be caught in the test.