ENH: Create meaningful broadcasting errors
Also checks dimensions a bit stricter (not relavent currently),
silence error (back to old behaviour...)

When advanced indexes cannot be broadcast, an index error is
now raised. When the assignment values cannot be broadcast
a ValueError is raised.

Note shape printing in nditer is no changed so that it is (2,)
and not (2).