ENH: Added argv argument to run_module_suite.
Adding an 'argv' optional argument to the
"numpy.testing.run_module_suite" function lets the client pass arguments
to the nose testrunner.

For example, this lets the user add things like:

run_module_suite(sys.argv)

at the end of a test module, so that arguments specified on the command
line will get passed to the nose runner. The user could then run:

$ python test_module.py --nocapture