BLD: remove "-Wl,-framework,Python" from linker flags for Intel compiler.
The "-Wl,-undefined,dynamic_lookup" flag is enough to make things work
with OS X bundles.  On the mailing list an issue was reported with
Anaconda (non-framework build) where f2py was picking up the system framework
Python instead of the Anaconda one.

Thanks to Robert Kern for explaining how to fix this.